Add files via upload
The tech layoff analysis project is a data analysis and visualization project that examines the trends and patterns in the layoff data of over 450 technology companies. The project involves using Python libraries such as Pandas, NumPy, Matplotlib, and Seaborn to perform tasks such as identifying the locations most impacted by layoffs, visualizing current layoffs trends based on months, analyzing the impact of IPO status on layoffs severity, identifying the industry with the most layoffs, and identifying the company with the most layoffs. The project aims to provide insights and recommendations based on the data analyzed. The results are presented in the form of tables, graphs, and charts, and the code and data are stored in a GitHub repository for easy access and collaboration.